deleteing ability has been fully feeded

+++ convex/files.ts
@@ -32,7 +32,8 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
     return null;
   }
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some((item) => item.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) {
     return null;
@@ -128,6 +129,14 @@ export const deleteFile = mutation({
 
     if (!access) {
       throw new ConvexError("you must be logged in to delete a file");
+    }
+
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError("only admin have access to delete a file");
     }
 
     await ctx.db.delete(args.fileId);
